Safari-611.1.10/LayoutTests/inspector/debugger/tail-deleted-frames/tail-deleted-frames-scopes-expected.txt

Ensure proper scopes in tail deleted frames.


== Running test suite: Debugger.TailDeletedFrames.Scopes
-- Running test case: Debugger.TailDeletedFrames.Scopes
PAUSED
CALL STACK:
0: [F] a
  ----Scope----
  x: 20
  -------------
1: -F- b
  ----Scope----
  y: 40
  -------------
2: -F- c
  ----Scope----
  z: 60
  -------------
3: [F] startABC
  ----Scope----
  --  empty  --
  -------------
4: [P] Global Code
  ----Scope----
  --  empty  --
  -------------
RESUMED
